Осипян Пайцар. Задание2

+++ task02/src/com/example/task02/Bill.java
@@ -34,7 +34,7 @@ public class Bill {
      *
      * @return общую стоимость покупки
      */
-    public long getPrice() {
+    public double getPrice() {
         long price = 0;
         for (BillItem item: items) {
             price += item.item.getPrice() * item.amount;
